back/requirements.txt
Django==5.0.6
mypy==1.10.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1